Only run post PotD action an hour early (#121)
Currently, the action runs twice:
1. an hour before the new problem is expected to be posted (11pm UTC)
2. when the new problem is expected to be posted (midnight UTC)

The first of these is the one that is actually doing all the posting. It won't have a new problem ready when it first runs, but it will sleep until it does.

By the time the second of these runs, the new problem of the day has already been posted, and so we try sleeping for the rest of the day. The action then ends up getting canceled by GitHub for running too long (the limit appears to be 6 hours).

So let's get rid of the second run, since the early run is working well.