Differentiate Python 3 goody name from its module name
The module name needs to be a valid Python 3 module name, whereas in the goody name it might be nice to support other characters. For example, I like calling the goody `int.digits` even though that's not a valid module name.

This also matches with how Java and Kotlin goodies already work.